Fix #242 Number input does not show validation errors if you enter te… (#247)
* Fix #242 Number input does not show validation errors if you enter text

The major browsers do not pass on a invalid number input as value to the javascript. Instead a empty value is passed. This blocks our validators from validating the invalid chars and creates a difference between what the browser displays and what stored as field value ( 'one' is stored as null).

As fix/work-around we can check the validity of the input ( supported by most, but not all browsers )  on keyup and if the inout is not valid put back the value as it was set before the input event.

* Test chrome local on CI WIP

* use both , and . as decimal separator in tests to cope with multiple locales

* work around for clearing fields on different browsers

* remove nightwatch locale setting and move invalid char test to separate test

* fix issue with incorrect id in example

* update geckodriver ( for firefox)

* use browser locale to determine decimal separator

* add work around for send keys as array issue

* And now for something completely different

* fix description and clear after input

* move empty test to avoid conflict on ff browser

* add clear value for remote ff issue

* use selenium backspace key code

* add unit test for onKeyUp method